Slew of issues 6-12-2017
Changed Specified overhangs and Overhang choice lists so that they actually make sense ("3'-ATCG-5' (Overhang)").
Added clauses for cases where the overhang choice is near the end of the repeat and there is blank space. Now prints 'N' for a random BP.
Fixed new off-by-one error created during the previous update so the last overhang prints correctly. 
Got rid of a backspace that was causing 'Oligos to order' list to look like trash in Notepad++.